PyMoose documentation (#1151)
* add docstrings to all edsl functions in pymoose

* docstrings throughout pymoose

* fix lint errors w/ new rust stable version

* bump pymoose dep versions

* bump onnx version

* re-do lint

+++ moose/src/fixedpoint/mod.rs
@@ -114,7 +114,7 @@ where
 
         // Exclude coefficients under precision
         for coeff in coeffs.iter().rev() {
-            if *coeff < 2f64.powi(-(x.fractional_precision() as i32 + 1)) as f64 {
+            if *coeff < 2f64.powi(-(x.fractional_precision() as i32 + 1)) {
                 degree -= 1
             } else {
                 break;
